asi_agents/requirements.txt
# Fetch.ai uAgents framework (requires Python 3.8+)
uagents

# Core dependencies for HTTP bridge and LLM integration
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.20.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# AI/LLM integration
google-generativeai>=0.3.0

# Web3 integration (optional - for blockchain features)
web3>=6.15.0

# Data handling
pydantic>=1.10.0
